api_gateway/requirements.txt
flask
flask-restful
grpcio-tools
Jinja2 
pytest
MarkupSafe
grpc-interceptor
flask_migrate
requests
flask-cors
grpcio